性能测试还差一个

+++ src/test/LogRecordTest.java
@@ -17,7 +17,7 @@ import org.junit.Test;
 public class LogRecordTest {
 
   private LogRecord constructLogRecord(int type, long txID, long key, byte[] value,
-      ArrayList<Long> activeTxns, long activeTxnStartEarlistOffset, int size) {
+      ArrayList<Long> activeTxns, int activeTxnStartEarlistOffset, int size) {
     LogRecord logRecord = new LogRecord();
     logRecord.setType(type);
     logRecord.setTxID(txID);
@@ -34,17 +34,17 @@ public class LogRecordTest {
   public void LogRecordTest1() {
     ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
     // start txn
-    LogRecord logRecord1 = constructLogRecord(LogRecordType.START_TXN, 1L, 0L, null, null, 0L, 13);
+    LogRecord logRecord1 = constructLogRecord(LogRecordType.START_TXN, 1L, 0L, null, null, 0, 13);
     byte[] b = logRecord1.getByteArray(outputStream);
     assert(logRecord1.equals(BytesUtils.byteToLogRecord(b)));
 
     // operate
-    LogRecord logRecord2 = constructLogRecord(LogRecordType.OPERATION, 1L, 1L, "hello world".getBytes(), null, 0L, 32);
+    LogRecord logRecord2 = constructLogRecord(LogRecordType.OPERATION, 1L, 1L, "hello world".getBytes(), null, 0, 32);
     b = logRecord2.getByteArray(outputStream);
     assert(logRecord2.equals(BytesUtils.byteToLogRecord(b)));
 
     // commit
-    LogRecord logRecord3 = constructLogRecord(LogRecordType.COMMIT_TXN, 1L, 0L, null, null, 0L, 13);
+    LogRecord logRecord3 = constructLogRecord(LogRecordType.COMMIT_TXN, 1L, 0L, null, null, 0, 13);
     b = logRecord3.getByteArray(outputStream);
     assert(logRecord3.equals(BytesUtils.byteToLogRecord(b)));
 
@@ -54,7 +54,7 @@ public class LogRecordTest {
     assert(logRecord4.equals(BytesUtils.byteToLogRecord(b)));
 
     // END_CKPT
-    LogRecord logRecord5 = constructLogRecord(LogRecordType.END_CKPT, 0L, 0L, null, null, 0L, 5);
+    LogRecord logRecord5 = constructLogRecord(LogRecordType.END_CKPT, 0L, 0L, null, null, 0, 5);
     b = logRecord5.getByteArray(outputStream);
     assert(logRecord5.equals(BytesUtils.byteToLogRecord(b)));
 
@@ -66,7 +66,7 @@ public class LogRecordTest {
     assert(logRecord6.equals(BytesUtils.byteToLogRecord(b)));
 
     // abort
-    LogRecord logRecord7 = constructLogRecord(LogRecordType.ABORT_TXN, 1L, 0L, null, null, 0L, 13);
+    LogRecord logRecord7 = constructLogRecord(LogRecordType.ABORT_TXN, 1L, 0L, null, null, 0, 13);
     b = logRecord7.getByteArray(outputStream);
     assert(logRecord7.equals(BytesUtils.byteToLogRecord(b)));
   }
